started pulling apart title screen code
