+++ src/frontend.cpp
@@ -79,6 +79,9 @@
 #include "map.h" //for builtInMap
 #include "notifications.h"
 
+#include "uiscreens/titlemenu.h"
+#include "uiscreens/common.h"
+
 // ////////////////////////////////////////////////////////////////////////////
 // Global variables
 
@@ -88,191 +91,6 @@ bool			bLimiterLoaded = false;
 
 #define TUTORIAL_LEVEL "TUTORIAL3"
 #define TRANSLATION_URL "https://translate.wz2100.net"
-
-// ////////////////////////////////////////////////////////////////////////////
-// Forward definitions
-
-static W_BUTTON * addSmallTextButton(UDWORD id, UDWORD PosX, UDWORD PosY, const char *txt, unsigned int style);
-static std::shared_ptr<W_BUTTON> makeTextButton(UDWORD id, const std::string &txt, unsigned int style);
-static std::shared_ptr<W_SLIDER> makeFESlider(UDWORD id, UDWORD parent, UDWORD stops, UDWORD pos);
-static std::shared_ptr<WIDGET> addMargin(std::shared_ptr<WIDGET> widget);
-
-// ////////////////////////////////////////////////////////////////////////////
-// Helper functions
-
-// Returns true if escape key pressed.
-//
-bool CancelPressed()
-{
-	const bool cancel = keyPressed(KEY_ESC);
-	if (cancel)
-	{
-		inputLoseFocus();	// clear the input buffer.
-	}
-	return cancel;
-}
-
-// Cycle through options as in program seq(1) from coreutils
-// The T cast is to cycle through enums.
-template <typename T> static T seqCycle(T value, T min, int inc, T max)
-{
-	return widgGetButtonKey_DEPRECATED(psWScreen) == WKEY_PRIMARY ?
-		value < max ? T(value + inc) : min :  // Cycle forwards.
-		min < value ? T(value - inc) : max;  // Cycle backwards.
-}
-
-// Cycle through options, which are powers of two, such as [128, 256, 512, 1024, 2048].
-template <typename T> static T pow2Cycle(T value, T min, T max)
-{
-	return widgGetButtonKey_DEPRECATED(psWScreen) == WKEY_PRIMARY ?
-		value < max ? std::max<T>(1, value) * 2 : min :  // Cycle forwards.
-		min < value ? (value / 2 > 1 ? value / 2 : 0) : max;  // Cycle backwards.
-}
-
-static void moveToParentRightEdge(WIDGET *widget, int32_t right)
-{
-	if (auto parent = widget->parent())
-	{
-		widget->move(parent->width() - (widget->width() + right), widget->y());
-	}
-}
-
-// ////////////////////////////////////////////////////////////////////////////
-// Title Screen
-void startTitleMenu()
-{
-	intRemoveReticule();
-
-	addBackdrop();
-	addTopForm(false);
-	addBottomForm();
-
-	addTextButton(FRONTEND_SINGLEPLAYER, FRONTEND_POS2X, FRONTEND_POS2Y, _("Single Player"), WBUT_TXTCENTRE);
-	addTextButton(FRONTEND_MULTIPLAYER, FRONTEND_POS3X, FRONTEND_POS3Y, _("Multi Player"), WBUT_TXTCENTRE);
-	addTextButton(FRONTEND_TUTORIAL, FRONTEND_POS4X, FRONTEND_POS4Y, _("Tutorial"), WBUT_TXTCENTRE);
-	addTextButton(FRONTEND_OPTIONS, FRONTEND_POS5X, FRONTEND_POS5Y, _("Options"), WBUT_TXTCENTRE);
-
-	// check whether video sequences are installed
-	if (PHYSFS_exists("sequences/devastation.ogg"))
-	{
-		addTextButton(FRONTEND_PLAYINTRO, FRONTEND_POS6X, FRONTEND_POS6Y, _("View Intro"), WBUT_TXTCENTRE);
-	}
-	else
-	{
-		addTextButton(FRONTEND_PLAYINTRO, FRONTEND_POS6X, FRONTEND_POS6Y, _("View Intro"), WBUT_TXTCENTRE | WBUT_DISABLE);
-		widgSetTip(psWScreen, FRONTEND_PLAYINTRO, _("Videos are missing, download them from http://wz2100.net"));
-	}
-
-	if (findLastSave())
-	{
-		addTextButton(FRONTEND_CONTINUE, FRONTEND_POS7X, FRONTEND_POS7Y, _("Continue Last Save"), WBUT_TXTCENTRE);
-	}
-	else
-	{
-		addTextButton(FRONTEND_CONTINUE, FRONTEND_POS7X, FRONTEND_POS7Y, _("Continue Last Save"), WBUT_TXTCENTRE | WBUT_DISABLE);
-		widgSetTip(psWScreen, FRONTEND_CONTINUE, _("No last save available"));
-	}
-	addTextButton(FRONTEND_QUIT, FRONTEND_POS8X, FRONTEND_POS8Y, _("Quit Game"), WBUT_TXTCENTRE);
-	addSideText(FRONTEND_SIDETEXT, FRONTEND_SIDEX, FRONTEND_SIDEY, _("MAIN MENU"));
-
-	addSmallTextButton(FRONTEND_HYPERLINK, FRONTEND_POS9X, FRONTEND_POS9Y, _("Official site: http://wz2100.net/"), 0);
-	widgSetTip(psWScreen, FRONTEND_HYPERLINK, _("Come visit the forums and all Warzone 2100 news! Click this link."));
-	W_BUTTON * pRightAlignedButton = addSmallTextButton(FRONTEND_DONATELINK, FRONTEND_POS9X + 360, FRONTEND_POS9Y, _("Donate: http://donations.wz2100.net/"), 0);
-	moveToParentRightEdge(pRightAlignedButton, 1);
-	widgSetTip(psWScreen, FRONTEND_DONATELINK, _("Help support the project with our server costs, Click this link."));
-	pRightAlignedButton = addSmallTextButton(FRONTEND_CHATLINK, FRONTEND_POS9X + 360, 0, _("Chat with players on Discord or IRC"), 0);
-	moveToParentRightEdge(pRightAlignedButton, 6);
-	widgSetTip(psWScreen, FRONTEND_CHATLINK, _("Connect to Discord or Freenode through webchat by clicking this link."));
-	addMultiBut(psWScreen, FRONTEND_BOTFORM, FRONTEND_UPGRDLINK, 7, 7, MULTIOP_BUTW, MULTIOP_BUTH, _("Check for a newer version"), IMAGE_GAMEVERSION, IMAGE_GAMEVERSION_HI, true);
-}
-
-static void runUpgrdHyperlink()
-{
-	std::string link = "https://wz2100.net/versioncheck/?ver=";
-	std::string version = version_getVersionString();
-	std::string versionStr;
-	for (char ch : version)
-	{
-		versionStr += (ch == ' ') ? '_' : ch;
-	}
-	link += urlEncode(versionStr.c_str());
-	openURLInBrowser(link.c_str());
-}
-
-static void runHyperlink()
-{
-	openURLInBrowser("https://wz2100.net/");
-}
-
-static void rundonatelink()
-{
-	openURLInBrowser("http://donations.wz2100.net/");
-}
-
-static void runchatlink()
-{
-	openURLInBrowser("https://wz2100.net/webchat/");
-}
-
-void runContinue()
-{
-	SPinit(lastSaveMP ? LEVEL_TYPE::SKIRMISH : LEVEL_TYPE::CAMPAIGN);
-	sstrcpy(saveGameName, lastSavePath);
-	bMultiPlayer = lastSaveMP;
-	setCampaignNumber(getCampaign(lastSavePath));
-}
-
-bool runTitleMenu()
-{
-	WidgetTriggers const &triggers = widgRunScreen(psWScreen);
-	unsigned id = triggers.empty() ? 0 : triggers.front().widget->id; // Just use first click here, since the next click could be on another menu.
-
-	switch (id)
-	{
-	case FRONTEND_QUIT:
-		changeTitleMode(QUIT);
-		break;
-	case FRONTEND_MULTIPLAYER:
-		changeTitleMode(MULTI);
-		break;
-	case FRONTEND_SINGLEPLAYER:
-		changeTitleMode(SINGLE);
-		break;
-	case FRONTEND_OPTIONS:
-		changeTitleMode(OPTIONS);
-		break;
-	case FRONTEND_TUTORIAL:
-		changeTitleMode(TUTORIAL);
-		break;
-	case FRONTEND_PLAYINTRO:
-		changeTitleMode(SHOWINTRO);
-		break;
-	case FRONTEND_HYPERLINK:
-		runHyperlink();
-		break;
-	case FRONTEND_UPGRDLINK:
-		runUpgrdHyperlink();
-		break;
-	case FRONTEND_DONATELINK:
-		rundonatelink();
-		break;
-	case FRONTEND_CHATLINK:
-		runchatlink();
-		break;
-	case FRONTEND_CONTINUE:
-		runContinue();
-		changeTitleMode(LOADSAVEGAME);
-		break;
-
-	default:
-		break;
-	}
-
-	widgDisplayScreen(psWScreen); // show the widgets currently running
-
-	return true;
-}
-
 
 // ////////////////////////////////////////////////////////////////////////////
 // Tutorial Menu
@@ -2150,157 +1968,6 @@ bool runGameOptionsMenu()
 	return true;
 }
 
-struct TitleBitmapCache {
-	WzText formattedVersionString;
-	WzText modListText;
-	WzText gfxBackend;
-};
-
-// ////////////////////////////////////////////////////////////////////////////
-// drawing functions
-
-// show a background picture (currently used for version and mods labels)
-static void displayTitleBitmap(WZ_DECL_UNUSED WIDGET *psWidget, WZ_DECL_UNUSED UDWORD xOffset, WZ_DECL_UNUSED UDWORD yOffset)
-{
-	char modListText[MAX_STR_LENGTH] = "";
-
-	if (!getModList().empty())
-	{
-		sstrcat(modListText, _("Mod: "));
-		sstrcat(modListText, getModList().c_str());
-	}
-
-	assert(psWidget->pUserData != nullptr);
-	TitleBitmapCache& cache = *static_cast<TitleBitmapCache*>(psWidget->pUserData);
-
-	cache.formattedVersionString.setText(version_getFormattedVersionString(), font_regular);
-	cache.modListText.setText(modListText, font_regular);
-	cache.gfxBackend.setText(gfx_api::context::get().getFormattedRendererInfoString(), font_small);
-
-	cache.formattedVersionString.render(pie_GetVideoBufferWidth() - 9, pie_GetVideoBufferHeight() - 14, WZCOL_GREY, 270.f);
-
-	if (!getModList().empty())
-	{
-		cache.modListText.render(9, 14, WZCOL_GREY);
-	}
-
-	cache.gfxBackend.render(9, pie_GetVideoBufferHeight() - 10, WZCOL_GREY);
-
-	cache.formattedVersionString.render(pie_GetVideoBufferWidth() - 10, pie_GetVideoBufferHeight() - 15, WZCOL_TEXT_BRIGHT, 270.f);
-
-	if (!getModList().empty())
-	{
-		cache.modListText.render(10, 15, WZCOL_TEXT_BRIGHT);
-	}
-
-	cache.gfxBackend.render(10, pie_GetVideoBufferHeight() - 11, WZCOL_TEXT_BRIGHT);
-}
-
-// ////////////////////////////////////////////////////////////////////////////
-// show warzone logo
-static void displayLogo(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset)
-{
-	iV_DrawImage2(WzString::fromUtf8("image_fe_logo.png"), xOffset + psWidget->x(), yOffset + psWidget->y(), psWidget->width(), psWidget->height());
-}
-
-
-// ////////////////////////////////////////////////////////////////////////////
-// show, well have a guess..
-static void displayBigSlider(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset)
-{
-	W_SLIDER *Slider = (W_SLIDER *)psWidget;
-	int x = xOffset + psWidget->x();
-	int y = yOffset + psWidget->y();
-
-	iV_DrawImage(IntImages, IMAGE_SLIDER_BIG, x + STAT_SLD_OX, y + STAT_SLD_OY);			// draw bdrop
-
-	int sx = (Slider->width() - 3 - Slider->barSize) * Slider->pos / Slider->numStops;  // determine pos.
-	iV_DrawImage(IntImages, IMAGE_SLIDER_BIGBUT, x + 3 + sx, y + 3);								//draw amount
-}
-
-// ////////////////////////////////////////////////////////////////////////////
-// show text written on its side.
-static void displayTextAt270(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset)
-{
-	SDWORD		fx, fy;
-	W_LABEL		*psLab;
-
-	psLab = (W_LABEL *)psWidget;
-
-	// Any widget using displayTextAt270 must have its pUserData initialized to a (DisplayTextOptionCache*)
-	assert(psWidget->pUserData != nullptr);
-	DisplayTextOptionCache& cache = *static_cast<DisplayTextOptionCache*>(psWidget->pUserData);
-
-	// TODO: Only works for single-line (not "formatted text") labels
-	cache.wzText.setText(psLab->getString().toUtf8(), font_large);
-
-	fx = xOffset + psWidget->x();
-	fy = yOffset + psWidget->y() + cache.wzText.width();
-
-	cache.wzText.render(fx + 2, fy + 2, WZCOL_GREY, 270.f);
-	cache.wzText.render(fx, fy, WZCOL_TEXT_BRIGHT, 270.f);
-}
-
-// ////////////////////////////////////////////////////////////////////////////
-// show a text option.
-void displayTextOption(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset)
-{
-	SDWORD			fx, fy, fw;
-	W_BUTTON		*psBut = (W_BUTTON *)psWidget;
-	bool			hilight = false;
-	bool			greyOut = psWidget->UserData || (psBut->getState() & WBUT_DISABLE); // if option is unavailable.
-
-	// Any widget using displayTextOption must have its pUserData initialized to a (DisplayTextOptionCache*)
-	assert(psWidget->pUserData != nullptr);
-	DisplayTextOptionCache& cache = *static_cast<DisplayTextOptionCache*>(psWidget->pUserData);
-
-	cache.wzText.setText(psBut->pText.toUtf8(), psBut->FontID);
-
-	if (psBut->isMouseOverWidget()) // if mouse is over text then hilight.
-	{
-		hilight = true;
-	}
-
-	fw = cache.wzText.width();
-	fy = yOffset + psWidget->y() + (psWidget->height() - cache.wzText.lineSize()) / 2 - cache.wzText.aboveBase();
-
-	if (psWidget->style & WBUT_TXTCENTRE)							//check for centering, calculate offset.
-	{
-		fx = xOffset + psWidget->x() + ((psWidget->width() - fw) / 2);
-	}
-	else
-	{
-		fx = xOffset + psWidget->x();
-	}
-
-	PIELIGHT colour;
-
-	if (greyOut)														// unavailable
-	{
-		colour = WZCOL_TEXT_DARK;
-	}
-	else															// available
-	{
-		if (hilight)													// hilight
-		{
-			colour = WZCOL_TEXT_BRIGHT;
-		}
-		else if (psWidget->id == FRONTEND_HYPERLINK || psWidget->id == FRONTEND_DONATELINK || psWidget->id == FRONTEND_CHATLINK) // special case for our hyperlink
-		{
-			colour = WZCOL_YELLOW;
-		}
-		else														// don't highlight
-		{
-			colour = WZCOL_TEXT_MEDIUM;
-		}
-	}
-
-	cache.wzText.render(fx, fy, colour);
-
-	return;
-}
-
-
 // ////////////////////////////////////////////////////////////////////////////
 // ////////////////////////////////////////////////////////////////////////////
 // ////////////////////////////////////////////////////////////////////////////
@@ -2440,7 +2107,7 @@ W_LABEL *addSideText(UDWORD id, UDWORD PosX, UDWORD PosY, const char *txt)
 }
 
 // ////////////////////////////////////////////////////////////////////////////
-static std::shared_ptr<W_BUTTON> makeTextButton(UDWORD id, const std::string &txt, unsigned int style)
+std::shared_ptr<W_BUTTON> makeTextButton(UDWORD id, const std::string &txt, unsigned int style)
 {
 	W_BUTINIT sButInit;
 
@@ -2486,99 +2153,7 @@ static std::shared_ptr<W_BUTTON> makeTextButton(UDWORD id, const std::string &tx
 	return button;
 }
 
-static std::shared_ptr<WIDGET> addMargin(std::shared_ptr<WIDGET> widget)
-{
-	return Margin(0, 20).wrap(widget);
-}
 
-void addTextButton(UDWORD id,  UDWORD PosX, UDWORD PosY, const std::string &txt, unsigned int style)
-{
-	auto button = makeTextButton(id, txt, style);
-	if (style & WBUT_TXTCENTRE)
-	{
-		button->setGeometry(PosX, PosY, FRONTEND_BUTWIDTH, button->height());
-	}
-	else
-	{
-		button->move(PosX + 35, PosY);
-	}
-
-	widgGetFromID(psWScreen, FRONTEND_BOTFORM)->attach(button);
-}
-
-W_BUTTON * addSmallTextButton(UDWORD id,  UDWORD PosX, UDWORD PosY, const char *txt, unsigned int style)
-{
-	W_BUTINIT sButInit;
-
-	sButInit.formID = FRONTEND_BOTFORM;
-	sButInit.id = id;
-	sButInit.x = (short)PosX;
-	sButInit.y = (short)PosY;
-
-	// Align
-	if (!(style & WBUT_TXTCENTRE))
-	{
-		sButInit.width = (uint16_t)(iV_GetTextWidth(txt, font_small)) + 4;
-		sButInit.x += 35;
-	}
-	else
-	{
-		sButInit.style |= WBUT_TXTCENTRE;
-		sButInit.width = FRONTEND_BUTWIDTH;
-	}
-
-	// Enable right clicks
-	if (style & WBUT_SECONDARY)
-	{
-		sButInit.style |= WBUT_SECONDARY;
-	}
-
-	sButInit.UserData = (style & WBUT_DISABLE); // store disable state
-	sButInit.pUserData = new DisplayTextOptionCache();
-	sButInit.onDelete = [](WIDGET *psWidget) {
-		assert(psWidget->pUserData != nullptr);
-		delete static_cast<DisplayTextOptionCache *>(psWidget->pUserData);
-		psWidget->pUserData = nullptr;
-	};
-
-	sButInit.height = FRONTEND_BUTHEIGHT;
-	sButInit.pDisplay = displayTextOption;
-	sButInit.FontID = font_small;
-	sButInit.pText = txt;
-	W_BUTTON * pButton = widgAddButton(psWScreen, &sButInit);
-
-	// Disable button
-	if (style & WBUT_DISABLE)
-	{
-		widgSetButtonState(psWScreen, id, WBUT_DISABLE);
-	}
-	return pButton;
-}
-
-// ////////////////////////////////////////////////////////////////////////////
-static std::shared_ptr<W_SLIDER> makeFESlider(UDWORD id, UDWORD parent, UDWORD stops, UDWORD pos)
-{
-	W_SLDINIT sSldInit;
-	sSldInit.formID		= parent;
-	sSldInit.id			= id;
-	sSldInit.width		= iV_GetImageWidth(IntImages, IMAGE_SLIDER_BIG);
-	sSldInit.height		= iV_GetImageHeight(IntImages, IMAGE_SLIDER_BIG);
-	sSldInit.numStops	= (UBYTE) stops;
-	sSldInit.barSize	= iV_GetImageHeight(IntImages, IMAGE_SLIDER_BIG);
-	sSldInit.pos		= (UBYTE) pos;
-	sSldInit.pDisplay	= displayBigSlider;
-	sSldInit.pCallback  = intUpdateQuantitySlider;
-
-	auto slider = std::make_shared<W_SLIDER>(&sSldInit);
-	return slider;
-}
-
-void addFESlider(UDWORD id, UDWORD parent, UDWORD x, UDWORD y, UDWORD stops, UDWORD pos)
-{
-	auto slider = makeFESlider(id, parent, stops, pos);
-	slider->move(x, y);
-	widgGetFromID(psWScreen, parent)->attach(slider);
-}
 
 // ////////////////////////////////////////////////////////////////////////////
 // Change Mode
