+++ src/uiscreens/titlemenu.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "common.h"
+
+
+void startTitleMenu();
+bool runTitleMenu();
+void runHyperlink();
